Export performance improvements

+++ shiftleft-utils/export.py
@@ -9,38 +9,53 @@ import time
 import urllib.parse
 
 import joern2sarif.lib.convert as convertLib
-import requests
+import httpx
 from json2xml import json2xml
 from rich.progress import Progress
 
 import config
 from common import extract_org_id, get_all_apps, get_dataflow, get_findings_url, headers
 
+console = Console(color_system="auto")
 
-def export_csv(app_list, findings_dict, report_file):
-    with open(report_file, "w", newline="") as csvfile:
+
+def export_csv(app_list, findings, report_file):
+    if not os.path.exists(report_file):
+        with open(report_file, "w", newline="") as csvfile:
+            reportwriter = csv.writer(
+                csvfile,
+                dialect="excel",
+                delimiter=",",
+                quotechar='"',
+                quoting=csv.QUOTE_MINIMAL,
+            )
+            reportwriter.writerow(
+                [
+                    "App",
+                    "App Group",
+                    "Finding ID",
+                    "Type",
+                    "Category",
+                    "OWASP Category",
+                    "Severity",
+                    "Source Method",
+                    "Sink Method",
+                    "Source File",
+                    "Version First Seen",
+                    "Scan First Seen",
+                    "Internal ID",
+                    "CVSS 3.1 Rating",
+                    "CVSS Score",
+                    "Reachability",
+                ]
+            )
+    with open(report_file, "a", newline="") as csvfile:
         reportwriter = csv.writer(
-            csvfile, delimiter=",", quotechar='"', quoting=csv.QUOTE_MINIMAL
-        )
-        reportwriter.writerow(
-            [
-                "App",
-                "App Group",
-                "Finding ID",
-                "Type",
-                "Category",
-                "OWASP Category",
-                "Severity",
-                "Source Method",
-                "Sink Method",
-                "Source File",
-                "Version First Seen",
-                "Scan First Seen",
-                "Internal ID",
-                "CVSS 3.1 Rating",
-                "CVSS Score",
-                "Reachability",
-            ]
+            csvfile,
+            dialect="excel",
+            delimiter=",",
+            quotechar='"',
+            quoting=csv.QUOTE_MINIMAL,
         )
         for app in app_list:
             app_id = app.get("id")
@@ -52,7 +67,6 @@ def export_csv(app_list, findings_dict, report_file):
                     if tag.get("key") == "group":
                         app_group = tag.get("value")
                         break
-            findings = findings_dict[app_name]
             source_method = ""
             sink_method = ""
             cvss_31_severity_rating = ""
@@ -147,20 +161,23 @@ def export_csv(app_list, findings_dict, report_file):
                                 cvss_score,
                                 "reachable"
                                 if afinding.get("related_findings", [])
-                                else "",
+                                else "N/A",
                             ]
                         )
 
 
-def get_all_findings(org_id, app_name, version):
+def get_all_findings(client, org_id, app_name, version):
     """Method to retrieve all findings"""
     findings_list = []
     findings_url = get_findings_url(org_id, app_name, version)
     page_available = True
     while page_available:
-        # print (findings_url)
-        r = requests.get(findings_url, headers=headers)
-        if r.ok:
+        try:
+            r = client.get(findings_url, headers=headers, timeout=config.timeout)
+        except httpx.ReadTimeout as e:
+            console.print(f"Unable to retrieve findings for {app_name}")
+            continue
+        if r.status_code == 200:
             raw_response = r.json()
             if raw_response and raw_response.get("response"):
                 response = raw_response.get("response")
@@ -187,8 +204,7 @@ def get_all_findings(org_id, app_name, version):
                     page_available = False
         else:
             page_available = False
-            print(f"Unable to retrieve findings for {app_name}")
-            print(r.status_code, r.json())
+            console.print(f"Unable to retrieve findings for {app_name}")
     return findings_list
 
 
@@ -217,128 +233,133 @@ def export_report(org_id, app_list, report_file, format):
             total=len(app_list),
             start=True,
         )
-        for app in app_list:
-            app_id = app.get("id")
-            app_name = app.get("name")
-            progress.update(task, description=f"Processing [bold]{app_name}[/bold]")
-            findings = get_all_findings(org_id, app_id, None)
-            file_category_set = set()
-            if format == "xml" or report_file.endswith(".xml"):
-                app_report_file = report_file.replace(".xml", "-" + app_id + ".xml")
-                with open(app_report_file, mode="w") as rp:
-                    xml_data = json2xml.Json2xml(findings).to_xml()
-                    if xml_data:
-                        rp.write(xml_data)
-                        progress.console.print(
-                            f"Findings report successfully exported to {app_report_file}"
-                        )
-            elif format == "sarif":
-                app_sarif_file = report_file.replace(".sarif", "-" + app_id + ".sarif")
-                app_json_file = app_sarif_file.replace(".sarif", ".json")
-                with open(app_json_file, mode="w") as rp:
-                    json.dump(
-                        {app_name: findings},
-                        rp,
-                        ensure_ascii=True,
-                        indent=None,
-                    )
-                    rp.flush()
-                convertLib.convert_file(
-                    "ng-sast",
-                    os.getenv("TOOL_ARGS", ""),
-                    work_dir,
-                    app_json_file,
-                    app_sarif_file,
-                    None,
-                )
-                progress.console.print(
-                    f"SARIF file successfully exported to {app_sarif_file}"
-                )
-                os.remove(app_json_file)
-            elif format == "sl":
-                with open(report_file, mode="w") as rp:
-                    for af in findings:
-                        details = af.get("details")
-                        title = af.get("title")
-                        # filename could be found either in file_locations or fileName
-                        filename = ""
-                        if details and details.get("file_locations"):
-                            file_locations = details.get("file_locations")
-                            if len(file_locations):
-                                filename = (
-                                    file_locations[0].split(":")[0].split("/")[-1]
-                                )
-                                filename = filename.replace(".java", "")
-                        # If there is no file_locations try to extract the name from the title
-                        if not filename and "BenchmarkTest" in title:
-                            filename = (
-                                title.split(" in ")[-1].replace("`", "").split(".")[0]
+        limits = httpx.Limits(max_keepalive_connections=5, max_connections=10)
+        with httpx.Client(http2=True, limits=limits) as client:
+            for app in app_list:
+                app_id = app.get("id")
+                app_name = app.get("name")
+                progress.update(task, description=f"Processing [bold]{app_name}[/bold]")
+                findings = get_all_findings(client, org_id, app_id, None)
+                file_category_set = set()
+                if format == "xml" or report_file.endswith(".xml"):
+                    app_report_file = report_file.replace(".xml", "-" + app_id + ".xml")
+                    with open(app_report_file, mode="w") as rp:
+                        xml_data = json2xml.Json2xml(findings).to_xml()
+                        if xml_data:
+                            rp.write(xml_data)
+                            progress.console.print(
+                                f"Findings report successfully exported to {app_report_file}"
                             )
-                        if filename.startswith("BenchmarkTest"):
-                            filename = filename.replace("BenchmarkTest", "")
-                        else:
-                            # Try to get the filename from source_method in details
-                            source_method = details.get("source_method")
-                            if source_method and "BenchmarkTest" in source_method:
-                                filename = source_method.split(":")[0].split(".")[4]
+                elif format == "sarif":
+                    app_sarif_file = report_file.replace(
+                        ".sarif", "-" + app_id + ".sarif"
+                    )
+                    app_json_file = app_sarif_file.replace(".sarif", ".json")
+                    with open(app_json_file, mode="w") as rp:
+                        json.dump(
+                            {app_name: findings},
+                            rp,
+                            ensure_ascii=True,
+                            indent=None,
+                        )
+                        rp.flush()
+                    convertLib.convert_file(
+                        "ng-sast",
+                        os.getenv("TOOL_ARGS", ""),
+                        work_dir,
+                        app_json_file,
+                        app_sarif_file,
+                        None,
+                    )
+                    progress.console.print(
+                        f"SARIF file successfully exported to {app_sarif_file}"
+                    )
+                    os.remove(app_json_file)
+                elif format == "sl":
+                    with open(report_file, mode="w") as rp:
+                        for af in findings:
+                            details = af.get("details")
+                            title = af.get("title")
+                            # filename could be found either in file_locations or fileName
+                            filename = ""
+                            if details and details.get("file_locations"):
+                                file_locations = details.get("file_locations")
+                                if len(file_locations):
+                                    filename = (
+                                        file_locations[0].split(":")[0].split("/")[-1]
+                                    )
+                                    filename = filename.replace(".java", "")
+                            # If there is no file_locations try to extract the name from the title
+                            if not filename and "BenchmarkTest" in title:
+                                filename = (
+                                    title.split(" in ")[-1]
+                                    .replace("`", "")
+                                    .split(".")[0]
+                                )
+                            if filename.startswith("BenchmarkTest"):
                                 filename = filename.replace("BenchmarkTest", "")
                             else:
+                                # Try to get the filename from source_method in details
+                                source_method = details.get("source_method")
+                                if source_method and "BenchmarkTest" in source_method:
+                                    filename = source_method.split(":")[0].split(".")[4]
+                                    filename = filename.replace("BenchmarkTest", "")
+                                else:
+                                    progress.console.print(
+                                        f'Get dataflow for {af.get("id")}'
+                                    )
+                                    # dataflows = get_dataflow(app_name, af.get("id"))
+                                    dataflows = details.get("dataflow", {}).get("list")
+                                    if dataflows:
+                                        for df in dataflows:
+                                            location = df.get("location")
+                                            if location.get(
+                                                "class_name"
+                                            ) and "BenchmarkTest" in location.get(
+                                                "class_name"
+                                            ):
+                                                filename = location.get(
+                                                    "class_name"
+                                                ).split(".")[-1]
+                                                filename = filename.replace(
+                                                    "BenchmarkTest", ""
+                                                )
+                                                break
+                            if not filename:
                                 progress.console.print(
-                                    f'Get dataflow for {af.get("id")}'
+                                    f"Unable to extract filename from file_locations or title {title}. Skipping ..."
                                 )
-                                # dataflows = get_dataflow(app_name, af.get("id"))
-                                dataflows = details.get("dataflow", {}).get("list")
-                                if dataflows:
-                                    for df in dataflows:
-                                        location = df.get("location")
-                                        if location.get(
-                                            "class_name"
-                                        ) and "BenchmarkTest" in location.get(
-                                            "class_name"
-                                        ):
-                                            filename = location.get("class_name").split(
-                                                "."
-                                            )[-1]
-                                            filename = filename.replace(
-                                                "BenchmarkTest", ""
-                                            )
-                                            break
-                        if not filename:
-                            progress.console.print(
-                                f"Unable to extract filename from file_locations or title {title}. Skipping ..."
+                                continue
+                            cwes = (
+                                int(pair["value"])
+                                for pair in af["tags"]
+                                if pair["key"] == "cwe_category"
                             )
-                            continue
-                        cwes = (
-                            int(pair["value"])
-                            for pair in af["tags"]
-                            if pair["key"] == "cwe_category"
+                            categories = (
+                                config.sl_owasp_category[cwe]
+                                for cwe in cwes
+                                if cwe in config.sl_owasp_category
+                            )
+                            try:
+                                category = next(categories)
+                                file_category = f"{filename},{category}"
+                                if file_category not in file_category_set:
+                                    rp.write(file_category + "\n")
+                                    file_category_set.add(file_category)
+                            except StopIteration:
+                                pass
+                        progress.console.print(
+                            f"Findings report successfully exported to {report_file}"
                         )
-                        categories = (
-                            config.sl_owasp_category[cwe]
-                            for cwe in cwes
-                            if cwe in config.sl_owasp_category
-                        )
-                        try:
-                            category = next(categories)
-                            file_category = f"{filename},{category}"
-                            if file_category not in file_category_set:
-                                rp.write(file_category + "\n")
-                                file_category_set.add(file_category)
-                        except StopIteration:
-                            pass
-                    progress.console.print(
-                        f"Findings report successfully exported to {report_file}"
-                    )
-            else:
-                findings_dict[app_name] = findings
-            progress.advance(task)
+                elif format == "csv":
+                    export_csv([app], findings, report_file)
+                else:
+                    findings_dict[app_name] = findings
+                progress.advance(task)
     if format == "json":
         with open(report_file, mode="w") as rp:
             json.dump(findings_dict, rp, ensure_ascii=True, indent=config.json_indent)
-            print(f"JSON report successfully exported to {report_file}")
-    if format == "csv":
-        export_csv(app_list, findings_dict, report_file)
-        print(f"CSV report successfully exported to {report_file}")
+            console.print(f"JSON report successfully exported to {report_file}")
 
 
 def build_args():
@@ -358,14 +379,14 @@ def build_args():
         "--report_file",
         dest="report_file",
         help="Report filename",
-        default="ngsast-report.json",
+        default="ngsast-report.csv",
     )
     parser.add_argument(
         "-f",
         "--format",
         dest="format",
         help="Report format",
-        default="json",
+        default="csv",
         choices=["json", "xml", "csv", "sl", "sarif"],
     )
     return parser.parse_args()
@@ -373,19 +394,19 @@ def build_args():
 
 if __name__ == "__main__":
     if not config.SHIFTLEFT_ACCESS_TOKEN:
-        print(
+        console.print(
             "Set the environment variable SHIFTLEFT_ACCESS_TOKEN before running this script"
         )
         sys.exit(1)
 
     org_id = extract_org_id(config.SHIFTLEFT_ACCESS_TOKEN)
     if not org_id:
-        print(
+        console.print(
             "Ensure the environment varibale SHIFTLEFT_ACCESS_TOKEN is copied exactly as-is from the website"
         )
         sys.exit(1)
 
-    print(config.ngsast_logo)
+    console.print(config.ngsast_logo)
     start_time = time.monotonic_ns()
     args = build_args()
     app_list = []
@@ -395,14 +416,12 @@ if __name__ == "__main__":
     format = args.format
     # Fix file extensions for xml format
     if format == "xml":
-        report_file = report_file.replace(".json", ".xml")
+        report_file = report_file.replace(".csv", ".xml")
     if format == "sarif":
-        report_file = report_file.replace(".json", ".sarif")
-    if format == "csv":
-        report_file = report_file.replace(".json", ".csv")
+        report_file = report_file.replace(".csv", ".sarif")
     elif format == "sl":
         if not args.app_name:
-            print(
+            console.print(
                 "This format is only suitable for OWASP Benchmark purposes. Use json or csv for all other apps"
             )
             sys.exit(1)
